modify EndSlaughter

+++ internal/app/handlers/request/slaughter.go
@@ -12,22 +12,22 @@ type ReqEndSlaughter struct {
 	Worker      string `json:"worker" form:"worker"`
 	HouseNumber string `json:"house_number" form:"house_number"`
 
-	SlaughterDisinfectHotWaterTempMoni1 float32 `json:"slaughter_disinfect_hot_water_temp_moni_1" form:"slaughter_disinfect_hot_water_temp_moni_1"` //预剥皮
-	SlaughterDisinfectHotWaterTempMoni2 float32 `json:"slaughter_disinfect_hot_water_temp_moni_2" form:"slaughter_disinfect_hot_water_temp_moni_2"` //检疫台
-	SlaughterDisinfectHotWaterTempMoni3 float32 `json:"slaughter_disinfect_hot_water_temp_moni_3" form:"slaughter_disinfect_hot_water_temp_moni_3"` //去皮后修正
-	SlaughterDisinfectHotWaterTempMoni4 float32 `json:"slaughter_disinfect_hot_water_temp_moni_4" form:"slaughter_disinfect_hot_water_temp_moni_4"` //去脏后
-	SlaughterDisinfectHotWaterTempMoni5 float32 `json:"slaughter_disinfect_hot_water_temp_moni_5" form:"slaughter_disinfect_hot_water_temp_moni_5"` //剔骨台
-	SlaughterDisinfectHotWaterTempMoni6 float32 `json:"slaughter_disinfect_hot_water_temp_moni_6" form:"slaughter_disinfect_hot_water_temp_moni_6"` //洗手台
+	SlaughterDisinfectHotWaterTempMoni1 float64 `json:"slaughter_disinfect_hot_water_temp_moni_1" form:"slaughter_disinfect_hot_water_temp_moni_1"` //预剥皮
+	SlaughterDisinfectHotWaterTempMoni2 float64 `json:"slaughter_disinfect_hot_water_temp_moni_2" form:"slaughter_disinfect_hot_water_temp_moni_2"` //检疫台
+	SlaughterDisinfectHotWaterTempMoni3 float64 `json:"slaughter_disinfect_hot_water_temp_moni_3" form:"slaughter_disinfect_hot_water_temp_moni_3"` //去皮后修正
+	SlaughterDisinfectHotWaterTempMoni4 float64 `json:"slaughter_disinfect_hot_water_temp_moni_4" form:"slaughter_disinfect_hot_water_temp_moni_4"` //去脏后
+	SlaughterDisinfectHotWaterTempMoni5 float64 `json:"slaughter_disinfect_hot_water_temp_moni_5" form:"slaughter_disinfect_hot_water_temp_moni_5"` //剔骨台
+	SlaughterDisinfectHotWaterTempMoni6 float64 `json:"slaughter_disinfect_hot_water_temp_moni_6" form:"slaughter_disinfect_hot_water_temp_moni_6"` //洗手台
 
-	Stun1 float32 `json:"stun_1" form:"stun_1"` //电压
-	Stun2 float32 `json:"stun_2" form:"stun_2"` //电流
-	Stun3 float32 `json:"stun_3" form:"stun_3"` //作用时间
+	Stun1 float64 `json:"stun_1" form:"stun_1"` //电压
+	Stun2 float64 `json:"stun_2" form:"stun_2"` //电流
+	Stun3 float64 `json:"stun_3" form:"stun_3"` //作用时间
 
-	BleedElectronic1 float32 `json:"bleed_electronic_1" form:"bleed_electronic_1"` //放血刀热水消毒
-	BleedElectronic2 float32 `json:"bleed_electronic_2" form:"bleed_electronic_2"` //电流
-	BleedElectronic3 float32 `json:"bleed_electronic_3" form:"bleed_electronic_3"` //作用时间
-	BleedElectronic4 float32 `json:"bleed_electronic_4" form:"bleed_electronic_4"` //电刺激参数
-	BleedElectronic5 float32 `json:"bleed_electronic_5" form:"bleed_electronic_5"` //电刺激时间
+	BleedElectronic1 float64 `json:"bleed_electronic_1" form:"bleed_electronic_1"` //放血刀热水消毒
+	BleedElectronic2 float64 `json:"bleed_electronic_2" form:"bleed_electronic_2"` //电流
+	BleedElectronic3 float64 `json:"bleed_electronic_3" form:"bleed_electronic_3"` //作用时间
+	BleedElectronic4 float64 `json:"bleed_electronic_4" form:"bleed_electronic_4"` //电刺激参数
+	BleedElectronic5 float64 `json:"bleed_electronic_5" form:"bleed_electronic_5"` //电刺激时间
 
 	PreSlaQuanPic1 string `json:"pre_sla_quan_pic_1" form:"pre_sla_quan_pic_1"` //咬肌照片
 	PreSlaQuanPic2 string `json:"pre_sla_quan_pic_2" form:"pre_sla_quan_pic_2"` //舌头
@@ -44,37 +44,37 @@ type ReqEndSlaughter struct {
 	SlaughterAnalAfterSlaQuanCar3 string `json:"slaughter_anal_after_sla_quan_car_3" form:"slaughter_anal_after_sla_quan_car_3"` //腹股沟淋巴
 	SlaughterAnalAfterSlaQuanCar4 string `json:"slaughter_anal_after_sla_quan_car_4" form:"slaughter_anal_after_sla_quan_car_4"` //整个胴体
 
-	AnalMeatPhMoni1 float32 `json:"anal_meat_ph_moni_1" form:"anal_meat_ph_moni_1"` //宰后0 min
-	AnalMeatPhMoni2 float32 `json:"anal_meat_ph_moni_2" form:"anal_meat_ph_moni_2"` //宰后45min
-	AnalMeatPhMoni3 float32 `json:"anal_meat_ph_moni_3" form:"anal_meat_ph_moni_3"` //劈半后胴体
-	AnalMeatPhMoni4 float32 `json:"anal_meat_ph_moni_4" form:"anal_meat_ph_moni_4"` //排酸过程中胴体
-	AnalMeatPhMoni5 float32 `json:"anal_meat_ph_moni_5" form:"anal_meat_ph_moni_5"` //冷鲜肉
+	AnalMeatPhMoni1 float64 `json:"anal_meat_ph_moni_1" form:"anal_meat_ph_moni_1"` //宰后0 min
+	AnalMeatPhMoni2 float64 `json:"anal_meat_ph_moni_2" form:"anal_meat_ph_moni_2"` //宰后45min
+	AnalMeatPhMoni3 float64 `json:"anal_meat_ph_moni_3" form:"anal_meat_ph_moni_3"` //劈半后胴体
+	AnalMeatPhMoni4 float64 `json:"anal_meat_ph_moni_4" form:"anal_meat_ph_moni_4"` //排酸过程中胴体
+	AnalMeatPhMoni5 float64 `json:"anal_meat_ph_moni_5" form:"anal_meat_ph_moni_5"` //冷鲜肉
 
-	AnalCutWeight1  float32 `json:"anal_cut_weight_1" form:"anal_cut_weight_1"`   //劈半后胴体 温度
-	AnalCutWeight2  float32 `json:"anal_cut_weight_2" form:"anal_cut_weight_2"`   //排酸过程中胴体 温度
-	AnalCutWeight3  float32 `json:"anal_cut_weight_3" form:"anal_cut_weight_3"`   //屠宰到入排酸库的时间
-	AnalCutWeight4  float32 `json:"anal_cut_weight_4" form:"anal_cut_weight_4"`   //屠宰到入排酸库的记录
-	AnalCutWeight5  float32 `json:"anal_cut_weight_5" form:"anal_cut_weight_5"`   //排酸库胴体间隙
-	AnalCutWeight6  float32 `json:"anal_cut_weight_6" form:"anal_cut_weight_6"`   //分割前胴体重量记录，称重时间
-	AnalCutWeight7  float32 `json:"anal_cut_weight_7" form:"anal_cut_weight_7"`   //分割前胴体重量记录，重量
-	AnalCutWeight8  float32 `json:"anal_cut_weight_8" form:"anal_cut_weight_8"`   //分割前胴体重量记录，酮体编号
-	AnalCutWeight9  float32 `json:"anal_cut_weight_9" form:"anal_cut_weight_9"`   //分割前胴体温度
-	AnalCutWeight10 float32 `json:"anal_cut_weight_10" form:"anal_cut_weight_10"` //分割刀温度
-	AnalCutWeight11 float32 `json:"anal_cut_weight_11" form:"anal_cut_weight_11"` //分割刀记录
-	AnalCutWeight12 float32 `json:"anal_cut_weight_12" form:"anal_cut_weight_12"` //温湿度监控
+	AnalCutWeight1  float64 `json:"anal_cut_weight_1" form:"anal_cut_weight_1"`   //劈半后胴体 温度
+	AnalCutWeight2  float64 `json:"anal_cut_weight_2" form:"anal_cut_weight_2"`   //排酸过程中胴体 温度
+	AnalCutWeight3  float64 `json:"anal_cut_weight_3" form:"anal_cut_weight_3"`   //屠宰到入排酸库的时间
+	AnalCutWeight4  float64 `json:"anal_cut_weight_4" form:"anal_cut_weight_4"`   //屠宰到入排酸库的记录
+	AnalCutWeight5  float64 `json:"anal_cut_weight_5" form:"anal_cut_weight_5"`   //排酸库胴体间隙
+	AnalCutWeight6  float64 `json:"anal_cut_weight_6" form:"anal_cut_weight_6"`   //分割前胴体重量记录，称重时间
+	AnalCutWeight7  float64 `json:"anal_cut_weight_7" form:"anal_cut_weight_7"`   //分割前胴体重量记录，重量
+	AnalCutWeight8  float64 `json:"anal_cut_weight_8" form:"anal_cut_weight_8"`   //分割前胴体重量记录，酮体编号
+	AnalCutWeight9  float64 `json:"anal_cut_weight_9" form:"anal_cut_weight_9"`   //分割前胴体温度
+	AnalCutWeight10 float64 `json:"anal_cut_weight_10" form:"anal_cut_weight_10"` //分割刀温度
+	AnalCutWeight11 float64 `json:"anal_cut_weight_11" form:"anal_cut_weight_11"` //分割刀记录
+	AnalCutWeight12 float64 `json:"anal_cut_weight_12" form:"anal_cut_weight_12"` //温湿度监控
 
-	ToNumGermMon1 float32 `json:"to_num_germ_mon_1" form:"to_num_germ_mon_1"` //喷淋后胴体
-	ToNumGermMon2 float32 `json:"to_num_germ_mon_2" form:"to_num_germ_mon_2"` //剔骨案板
-	ToNumGermMon3 float32 `json:"to_num_germ_mon_3" form:"to_num_germ_mon_3"` //传送带
-	ToNumGermMon4 float32 `json:"to_num_germ_mon_4" form:"to_num_germ_mon_4"` //围裙
-	ToNumGermMon5 float32 `json:"to_num_germ_mon_5" form:"to_num_germ_mon_5"` //手套
-	ToNumGermMon6 float32 `json:"to_num_germ_mon_6" form:"to_num_germ_mon_6"` //锯骨机
-	ToNumGermMon7 float32 `json:"to_num_germ_mon_7" form:"to_num_germ_mon_7"` //刀具
-	ToNumGermMon8 float32 `json:"to_num_germ_mon_8" form:"to_num_germ_mon_8"` //地面与墙面
+	ToNumGermMon1 float64 `json:"to_num_germ_mon_1" form:"to_num_germ_mon_1"` //喷淋后胴体
+	ToNumGermMon2 float64 `json:"to_num_germ_mon_2" form:"to_num_germ_mon_2"` //剔骨案板
+	ToNumGermMon3 float64 `json:"to_num_germ_mon_3" form:"to_num_germ_mon_3"` //传送带
+	ToNumGermMon4 float64 `json:"to_num_germ_mon_4" form:"to_num_germ_mon_4"` //围裙
+	ToNumGermMon5 float64 `json:"to_num_germ_mon_5" form:"to_num_germ_mon_5"` //手套
+	ToNumGermMon6 float64 `json:"to_num_germ_mon_6" form:"to_num_germ_mon_6"` //锯骨机
+	ToNumGermMon7 float64 `json:"to_num_germ_mon_7" form:"to_num_germ_mon_7"` //刀具
+	ToNumGermMon8 float64 `json:"to_num_germ_mon_8" form:"to_num_germ_mon_8"` //地面与墙面
 
-	AirNumGermMon1 float32 `json:"air_num_germ_mon_1" form:"air_num_germ_mon_1"` //屠宰车间
-	AirNumGermMon2 float32 `json:"air_num_germ_mon_2" form:"air_num_germ_mon_2"` //分割车间
-	AirNumGermMon3 float32 `json:"air_num_germ_mon_3" form:"air_num_germ_mon_3"` //排酸车间
+	AirNumGermMon1 float64 `json:"air_num_germ_mon_1" form:"air_num_germ_mon_1"` //屠宰车间
+	AirNumGermMon2 float64 `json:"air_num_germ_mon_2" form:"air_num_germ_mon_2"` //分割车间
+	AirNumGermMon3 float64 `json:"air_num_germ_mon_3" form:"air_num_germ_mon_3"` //排酸车间
 
 	PreSlaDietManage1 string `json:"pre_sla_diet_manage_1" form:"pre_sla_diet_manage_1"` //牛进场时间
 	PreSlaDietManage2 string `json:"pre_sla_diet_manage_2" form:"pre_sla_diet_manage_2"` //牛停水时间
